requirements.txt
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
google-api-core==1.20.0
google-api-python-client==1.9.1
google-auth==1.16.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.25.9
wikipedia==1.4.0
